Moving _make_divisible to utils.
